internal/chip8: start work on emulating a cpu cycle
Add an `EmulateCycle` func which set’s the vm’s opcode, set’s our draw flag to false, then calls another new func `parseOpcode`. This func is unfinished, but serves rather as a map of all the work it will need to handle. All cases for opcodes described with notes on what needs to happen.